# api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NFC Loyalty & Customer Growth Platform API",
    "description": "REST API untuk platform loyalty berbasis NFC/QR untuk UMKM Indonesia. Menyediakan fitur registrasi member, pemesanan, validasi pembayaran manual, poin loyalitas, voucher, dan dashboard analytics.",
    "version": "1.0.0",
    "contact": { "name": "NFC Loyalty Platform" }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development server" }
  ],
...
  "paths": {
...
    "/api/vouchers/validate": {
      "post": {
        "tags": ["Vouchers"],
        "summary": "Validate voucher code (member)",
        "operationId": "validateVoucher",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code"],
                "properties": { "code": { "type": "string", "maxLength": 20 } }
              },
              "example": { "code": "DISKON10" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Voucher is valid",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "code": "DISKON10",
                    "discountType": "percentage",
                    "discountValue": 10,
                    "expiryDate": "2025-12-31T23:59:59.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid voucher",
            "content": {
              "application/json": {
                "examples": {
                  "invalid": { "summary": "Code not found", "value": { "success": false, "error": { "message": "Kode voucher tidak valid", "code": "INVALID_CODE" } } },
                  "expired": { "summary": "Voucher expired", "value": { "success": false, "error": { "message": "Voucher sudah expired", "code": "EXPIRED" } } },
                  "maxed": { "summary": "Usage limit reached", "value": { "success": false, "error": { "message": "Voucher sudah mencapai batas penggunaan", "code": "USAGE_LIMIT_REACHED" } } }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/vouchers": {
      "post": {
        "tags": ["Vouchers"],
        "summary": "Create voucher (admin)",
        "operationId": "createVoucher",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateVoucherRequest" },
              "example": {
                "code": "PROMO20",
                "discountType": "percentage",
                "discountValue": 20,
                "expiryDate": "2025-12-31T23:59:59.000Z",
                "maxUsage": 100
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Voucher created",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "v-001",
                    "tenantId": "550e8400-e29b-41d4-a716-446655440000",
                    "code": "PROMO20",
                    "discountType": "percentage",
                    "discountValue": 20,
                    "expiryDate": "2025-12-31T23:59:59.000Z",
                    "maxUsage": 100,
                    "currentUsage": 0,
                    "isActive": true,
                    "isWelcomeVoucher": false,
                    "createdAt": "2025-07-15T10:30:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Duplicate code or validation error",
            "content": {
              "application/json": {
                "example": { "success": false, "error": { "message": "Kode voucher sudah digunakan" } }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Vouchers"],
        "summary": "List vouchers with stats (admin)",
        "operationId": "listVouchers",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Voucher list with statistics",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "v-001",
                      "code": "PROMO20",
                      "discountType": "percentage",
                      "discountValue": 20,
                      "expiryDate": "2025-12-31T23:59:59.000Z",
                      "maxUsage": 100,
                      "currentUsage": 35,
                      "isActive": true,
                      "remainingUses": 65,
                      "redemptionRate": 35.0
                    },
                    {
                      "id": "v-002",
                      "code": "HEMAT5K",
                      "discountType": "fixed",
                      "discountValue": 5000,
                      "expiryDate": "2025-09-30T23:59:59.000Z",
                      "maxUsage": 50,
                      "currentUsage": 50,
                      "isActive": true,
                      "remainingUses": 0,
                      "redemptionRate": 100.0
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/vouchers/{id}/deactivate": {
      "patch": {
        "tags": ["Vouchers"],
        "summary": "Deactivate voucher (admin)",
        "operationId": "deactivateVoucher",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Voucher deactivated",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "v-001",
                    "code": "PROMO20",
                    "isActive": false
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token (15min). Get from /api/auth/register, /api/auth/login, or /api/auth/admin/login"
      }
    },
    "schemas": {
...
      "CreateVoucherRequest": {
        "type": "object",
        "required": ["code", "discountType", "discountValue", "expiryDate", "maxUsage"],
        "properties": {
          "code": { "type": "string", "minLength": 1, "maxLength": 20, "pattern": "^[a-zA-Z0-9]+$" },
          "discountType": { "type": "string", "enum": ["percentage", "fixed"] },
          "discountValue": { "type": "integer", "description": "1-100 for percentage, 1000-10000000 for fixed" },
          "expiryDate": { "type": "string", "format": "date-time", "description": "Must be future date (ISO 8601)" },
          "maxUsage": { "type": "integer", "minimum": 1 }
        }
      },
...
    }
  }
}